// web/chanlun_demo/app/static/charting_library/datafeed-api.d.ts
export interface Bar {
	/** Bar time.
	 * Amount of **milliseconds** since Unix epoch start in **UTC** timezone.
	 * `time` for daily, weekly, and monthly bars is expected to be a trading day (not session start day) at 00:00 UTC.
	 * The library adjusts time according to `session` from {@link LibrarySymbolInfo}.
	 */
	time: number;
	/** Opening price */
	open: number;
	/** High price */
	high: number;
	/** Low price */
	low: number;
	/** Closing price */
	close: number;
	/** Trading Volume */
	volume?: number;
}
// ...
export interface HistoryMetadata {
	/**
	 * Optional value indicating to the library that there is no more data on the server.
	 */
	noData?: boolean;
	/**
	 * The time of the next available bar in history. The time value should be represented with a Unix timestamp in milliseconds.
	 *
	 * You can pass the `nextTime` value to the library if there is no data in the time range that the library requests.
	 * Therefore, you notify the library about available data before the requested range and ensure that the next data request is for the right range.
	 * For more information, refer to the [How nextTime works](https://www.tradingview.com/charting-library-docs/latest/connecting_data/UDF#how-nexttime-works) section.
	 */
	nextTime?: number | null;
}
// ...
export interface PeriodParams {
	/**
	 * Unix timestamp (leftmost requested bar)
	 */
	from: number;
	/**
	 * Unix timestamp (rightmost requested bar - not inclusive)
	 */
	to: number;
	/**
	 * The exact amount of bars to load, should be considered a higher priority than `from` if your datafeed supports it
	 */
	countBack: number;
	/**
	 * Used to identify if it's the first call of getBars
	 */
	firstDataRequest: boolean;
}